results/experiment/llama_10_turns/scenario_5.5/prompt_2/requirements.txt
mysql_connector_repackaged==0.3.1
python_bcrypt==0.3.2
